update.1
Removido arquivo de routes;
Adicionado rotas no index.js;
Corrigido erro da varialvel targetBrand na função buscarMarca;
