--- routes/accounts.js
@@ -1,24 +1,0 @@
-import express from "express";
-const router = express.Router();
-/*
-app.get("/marcas/maisModelos", (req, res) => {
-  // Rota para o item 1
-});
-
-app.get("/marcas/menosModelos", (req, res) => {
-  // Rota para o item 2
-});
-
-app.get("/marcas/listaMaisModelos/:X", (req, res) => {
-  // Rota para o item 3
-});
-
-app.get("/marcas/listaMenosModelos/:X", (req, res) => {
-  // Rota para o item 4
-});
-
-app.post("/marcas/listaModelos", (req, res) => {
-  // Rota para o item 5
-});
-*/
-export default router;
